refactoring, tidy up and type checking
