+++ staresc/exporter/__init__.py
@@ -4,5 +4,10 @@ This module is responsible for the output of Staresc
 """
 
 from .exporter import Exporter
-from .handlers import Handler, CSVHandler, ScanHandler, XLSXHandler, JSONHandler, RawHandler
+from .handler import Handler
+from .csv_handler import CSVHandler
+from .json_handler import JSONHandler
+from .raw_handler import RawHandler
+from .xlsx_handler import XLSXHandler
+from .scan_handler import ScanHandler
